secret key functionality

+++ src/site-components/DeletePostButton.jsx
@@ -10,11 +10,22 @@ import {
     AlertDialogTrigger,
 } from "@/components/ui/alert-dialog";
 import { supabase } from "../client";
+import { Input } from "@/components/ui/input";
+import { useState } from "react";
 
-const DeletePostButton = ({ postid, pollid, btnCSS }) => {
+const DeletePostButton = ({ postid, pollid, btnCSS, password }) => {
+  const [inputKey, setInputKey] = useState("");
+  const [failedAttempt, setFailedAttempt] = useState(false);
+
     const deletePost = async (event) => {
         event.preventDefault();
-    
+
+        if (inputKey !== password) {
+          setFailedAttempt(true);
+          return;
+        }
+        setFailedAttempt(false);
+
         // Delete post
         await supabase.from("posts").delete().eq("id", postid);
     
@@ -28,6 +39,11 @@ const DeletePostButton = ({ postid, pollid, btnCSS }) => {
         window.location = "/forum";
     };
 
+    const handleChange = (event) => {
+      setFailedAttempt(false);
+      setInputKey(event.target.value);
+    };  
+
  return (
     <AlertDialog>
       <AlertDialogTrigger className={btnCSS === true ? "bg-black hover:bg-gray-800 text-white font-bold py-2 px-4 border-b-4 border-black rounded" : ""}>
@@ -38,9 +54,13 @@ const DeletePostButton = ({ postid, pollid, btnCSS }) => {
         <AlertDialogTitle>
           Are you sure?
         </AlertDialogTitle>
-        <AlertDialogDescription>This action cannot be undone. This will permanently delete the post.</AlertDialogDescription>
+        <AlertDialogDescription>This action cannot be undone. This will permanently delete the post. Enter the secret key: </AlertDialogDescription>
+        <Input 
+          onChange={handleChange}
+        />
       </AlertDialogHeader>
-      <AlertDialogFooter>
+      <AlertDialogFooter className="flex items-center">
+        <h3 className="text-red-500">{failedAttempt ?" Wrong passcode." : ""}</h3>
         <AlertDialogCancel>Cancel</AlertDialogCancel>
         <AlertDialogAction onClick={deletePost}>Delete Post</AlertDialogAction>
       </AlertDialogFooter>
